requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.9

# Database & Auth
SQLAlchemy==2.0.41
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Gemini AI
google-generativeai==0.8.5

# LangChain (RAG essentials)
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langchain-huggingface==0.1.0  # For embeddings

# Vector Store
faiss-cpu==1.11.0

# Document Processing
pypdf==5.5.0
python-docx==1.2.0

# Text/HTML
markdown==3.8
beautifulsoup4==4.13.4

# Utilities
pydantic==2.11.4
PyYAML==6.0.2
httpx==0.28.1